:art: Improve enum usage in bucket type and region
To make scalability faster, go-enum tool has been used to
manage the enums created in the application. It auto bundles
generation of parsing interfaces for smooth usage within the
code. Further, the bucket regions have been renamed to not
contain any region specific city name and instead follow a
similar region naming pattern as AWS. This allows the
underlying storage backends to match such region to nearest
available region that they support and there by minimising
the need for a major refactoring in the future due to issues